refactor(journey): decompõe SendMessagePanel em 3 componentes (CRM-126)
1060 linhas num arquivo só (o maior nó de jornada) — sem separação entre configuração de canal, corpo da mensagem (texto/template) e anexos, cada seção só existia por indentação. Molde: send-webhook, já decomposto em components/ há alguns épicos.

Painel caiu para 579 linhas — o que sobrou é estado, handlers e validação, não mais marcação. Extraído:
- SendMessageChannelConfig — toggle de modo, canal de evento, seleção de inbox.
- SendMessageContent — corpo da mensagem: texto livre OU template (picker, preview, mapeamento de variável por fonte), mutuamente exclusivos, um componente só decide qual renderizar em vez do painel ramificar em dois lugares.
- SendMessageAttachments — dropzone de drag-and-drop + lista de anexados.

Extração pura de JSX — todo estado (formData, attachments, templates) e handlers continuam no painel; os subcomponentes recebem os dados e os callbacks já existentes, sem reescrever nenhuma lógica de atualização. Os 12 testes do painel (que já exercitam o payload final ponta a ponta) continuaram verdes sem alteração, confirmando que o comportamento não mudou. Adicionado spec próprio para SendMessageContent (o pedaço mais arriscado) cobrindo os dois modos.
